FIX: design and content issues
